# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Mutant detector API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/mutant": {
            "post": {
                "description": "return true or false by dna input",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Check if given dna is from a mutant",
                "parameters": [
                    {
                        "description": "dna data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.IsMutantReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apperrors.DeliveryError"
                        }
                    }
                }
            }
        },
        "/stats": {
            "post": {
                "description": "return db stats about dna",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "returns db stats about dna",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dna.DnaReportStats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.DeliveryError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apperrors.DeliveryError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "dna.DnaReportStats": {
            "type": "object",
            "properties": {
                "count_human_dna": {
                    "type": "integer"
                },
                "count_mutant_dna": {
                    "type": "integer"
                },
                "ratio": {
                    "type": "number"
                }
            }
        },
        "models.IsMutantReqBody": {
            "type": "object",
            "required": [
                "dna"
            ],
            "properties": {
                "dna": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}